tmux: add keybinding to use some of the bash shortcuts
We are using prefix as Ctrl+A which conflicts with the bash key bindings
to got to start of command. We have changed the keybindings in the
previous commits to use <prefix> <C-a> for moving the cursor to the start.

We also define key maps for clearing the command from the cursor to the
end. The key binding is <prefix> + <C-k>

Signed-off-by: Tushar Jagad <tushar.jagad@gmail.com>