gtk: add css for gtk-3 and gtk-4
This patch add custom css for gtk-3 and gtk-4 to configure the window
frame and decoration. The problem was observed when using ghostty.
Similar observation was posted on reditt. The link to the same post is
https://www.reddit.com/r/awesomewm/comments/u3237d/comment/i8p75c9/

This has not helped getting rid of the double line around the ghostty
window but retaining these changes for now.

Signed-off-by: Tushar Jagad <tushar.jagad@gmail.com>