sway: config: update keybinding for audio sink toggle
This commit changes the keymap for toggling the audio sink in the Sway.

The path to the script has been updated from a relative path to an
absolute path using the HOME environment variable. This ensures that
the script is correctly located and executed regardless of the current
working directory.

Signed-off-by: Tushar Jagad <tushar.jagad@gmail.com>